// src/shared/model/users_list.model.ts
export class UsersModel {
    Id:string;
    FirebaseKey: string;
    FirstName: string;
    LastName: string;
    ClubKey: string;
    IsChild: string;
    DOB: string;
    EmailID: string;
    EmergencyContactName: string;
    EmergencyNumber: string;
    Gender: string;
    MedicalCondition: string;
    ParentClubKey: string;
    ParentKey: string;
    PhoneNumber: string;
    IsEnable: string;
    IsActive: string;
    //Source
  }